try to fix chat 8
