+++ src/socket.ts
@@ -2,4 +2,4 @@
 
 import { io } from 'socket.io-client';
 
-export const socket = io(process.env.NEXT_PUBLIC_CHAT_URL);
+export const socket = io();
